finished place model and updated the add-models middleware

+++ src/db/models/place.js
@@ -3,7 +3,7 @@ const knex = require("../knex");
 //Define CRUD interface
 class Place {
   //add a place to the favorites table
-  static async addFavorite(
+  static async addPlace(
     biz_id,
     name,
     address,
@@ -12,11 +12,13 @@ class Place {
     number,
     price_level,
     rating,
-    user_id
+    user_id,
+    emoji_rating,
+    is_favorited
   ) {
     try {
       const result = await knex.raw(
-        `INSERT INTO favorites (   
+        `INSERT INTO places (   
         biz_id,
         name,
         address,
@@ -25,9 +27,11 @@ class Place {
         number,
         price_level,
         rating,
-        user_id) 
+        user_id,
+        emoji_rating, 
+        is_favorited) 
         VALUES 
-        (?, ?, ?, ?, ?, ?, ?, ?, ?)
+        (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
         RETURNING *`,
         [
           biz_id,
@@ -39,9 +43,12 @@ class Place {
           price_level,
           rating,
           user_id,
+          emoji_rating,
+          is_favorited,
         ]
       );
       console.log(result);
+      return result.rows;
     } catch (err) {
       console.error(err);
       return null;
@@ -49,11 +56,32 @@ class Place {
   }
 
   //list a place
-  static async listPlace() {
+  static async listPlace(id) {
     try {
-      const result = await knew.raw(`
-      
-      `);
+      const result = await knex.raw(
+        `
+      SELECT * FROM places WHERE id=?
+      `,
+        [id]
+      );
+      console.log(result);
+      return result;
+    } catch (err) {
+      console.error(err);
+      return null;
+    }
+  }
+
+  static async listFavorite(favorite_id) {
+    try {
+      const result = await knex.raw(
+        `
+      SELECT * FROM favorites where id=?
+      `,
+        [favorite_id]
+      );
+      console.log(result.rows);
+      return result.rows;
     } catch (err) {
       console.error(err);
       return null;
@@ -75,11 +103,16 @@ class Place {
   }
 
   //remove/delete a favorite
-  static async removeFavorite(favorite_id) {
+  static async removeFavorite(favorite_id, user_id) {
     try {
-      const result = await knex.raw(`
-      DELETE FROM favorites where id=?
-      `)
+      const result = await knex.raw(
+        `
+      DELETE FROM favorites where id=? AND user_id=?
+      `,
+        [favorite_id, user_id]
+      );
+      console.log(result.rows);
+      return result.rows;
     } catch (err) {
       console.error(err);
       return null;
@@ -87,10 +120,15 @@ class Place {
   }
 }
 
-const testQueries = async () => {
-  const favorite = await Place.addFavorite();
-  console.log(favorite);
-};
-testQueries();
+// const testQueries = async () => {
+//   const favorite = await Place.addPlace(1, 2, 3, 4, 5, 6, 7, 8, 1, 10, false);
+//   console.log(favorite);
+//   const listPlace = await Place.listPlace(8);
+//   console.log("im listing the place", listPlace);
+
+//   const listingAllFavorites = await Place.listAllFavorites(1);
+//   console.log("im listing favorites", listingAllFavorites);
+// };
+// testQueries();
 
 module.exports = Place;
